ntp_proto: backport bugfix of 2952
In ntp 4.2.8p4, orphaned mode and other peer-type synchronizations were
broken. See the bug thread at:
http://bugs.ntp.org/show_bug.cgi?id=2952

It was fixed in ntp commit 19d58f66c96a9:
https://github.com/ntp-project/ntp/commit/19d58f66c96a

That commit is full of whitespace and logging fixes, so I picked it,
amended out everything unnecessary, and this is what we are left with.

It also conflicted with one of the backported ubuntu patches, but I
resolved it using the latest NTP code, so the resulting ifs are exactly
the same as the current upstream code.

References: jira PL-321